# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ledplayr",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.3.2"
  },
  "paths": {
    "/api/channel/output/universeOutputs": {
      "get": {
        "tags": [
          "FPP Compatibility"
        ],
        "summary": "Retrieve outputs.json",
        "description": "Download the outputs in JSON format",
        "operationId": "get_outputs",
        "responses": {
          "200": {
            "description": "Ouputs downloaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channels"
                }
              }
            }
          },
          "404": {
            "description": "Outputs not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FPP Compatibility"
        ],
        "summary": "Upload outputs.json",
        "description": "Upload the outputs in JSON format",
        "operationId": "upload_outputs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Channels"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Outputs uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/configfile/virtualdisplaymap": {
      "get": {
        "tags": [
          "FPP Compatibility"
        ],
        "summary": "Retrieve VirtualDisplayMap",
        "description": "Download the VirtualDisplayMap",
        "operationId": "get_display",
        "responses": {
          "200": {
            "description": "VirtualDisplayMap downloaded successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "VirtualDisplayMap not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FPP Compatibility"
        ],
        "summary": "Upload VirtualDisplayMap",
        "description": "Upload the VirtualDisplayMap",
        "operationId": "upload_display",
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "VDM uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models": {
      "get": {
        "tags": [
          "FPP Compatibility"
        ],
        "summary": "Retrieve models.json",
        "description": "Download the models in JSON format",
        "operationId": "list_models",
        "responses": {
          "200": {
            "description": "Models downloaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Model"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Models not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FPP Compatibility"
        ],
        "summary": "Upload models.json",
        "description": "Upload the models in JSON format",
        "operationId": "upload_models",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Models"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Models uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
...
    "/fppxml.php": {
      "get": {
        "tags": [
          "FPP Compatibility"
        ],
        "summary": "Run an FPP Command",
        "description": "This method isn't really implemented. The only command you\ncan issue is `moveFile` and all that really does is check\nwhether a file exists or not - it doesn't move it because\nthat's handled at upload time.",
        "operationId": "fpp_command",
        "parameters": [
          {
            "name": "command",
            "in": "query",
            "description": "The FFP command",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "moveFile"
          },
          {
            "name": "file",
            "in": "query",
            "description": "The file to move",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "file.fseq"
          }
        ],
        "responses": {
          "200": {
            "description": "File exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "400": {
            "description": "Unrecognized command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "404": {
            "description": "File doesn't exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChannelOutput": {
        "type": "object",
        "required": [
          "type",
          "startChannel",
          "enabled",
          "timeout",
          "channelCount",
          "universes"
        ],
        "properties": {
          "channelCount": {
            "type": "integer",
            "format": "int32"
          },
          "enabled": {
            "type": "boolean"
          },
          "startChannel": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "timeout": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "type": {
            "type": "string"
          },
          "universes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Universe"
            }
          }
        }
      },
      "Channels": {
        "type": "object",
        "required": [
          "channelOutputs"
        ],
        "properties": {
          "channelOutputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelOutput"
            }
          }
        }
      },
...
      "Model": {
        "type": "object",
        "required": [
          "Name",
          "xLights",
          "ChannelCount",
          "Orientation",
          "StartChannel",
          "StringCount",
          "ChannelCountPerNode",
          "StrandsPerString",
          "StartCorner",
          "Type"
        ],
        "properties": {
          "ChannelCount": {
            "type": "integer",
            "format": "int32",
            "example": 6,
            "minimum": 0
          },
          "ChannelCountPerNode": {
            "type": "integer",
            "format": "int32",
            "example": 3,
            "minimum": 0
          },
          "Name": {
            "type": "string",
            "example": "Single_line"
          },
          "Orientation": {
            "type": "string",
            "example": "horizontal"
          },
          "StartChannel": {
            "type": "integer",
            "format": "int32",
            "example": 1,
            "minimum": 0
          },
          "StartCorner": {
            "type": "string",
            "example": "BL"
          },
          "StrandsPerString": {
            "type": "integer",
            "format": "int32",
            "example": 1,
            "minimum": 0
          },
          "StringCount": {
            "type": "integer",
            "format": "int32",
            "example": 1,
            "minimum": 0
          },
          "Type": {
            "type": "string",
            "example": "Channel"
          },
          "xLights": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Models": {
        "type": "object",
        "required": [
          "models"
        ],
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          }
        }
      },
...
      "Status": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong",
            "example": "Could not open file",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "error"
          }
        }
      },
...
      "Universe": {
        "type": "object",
        "required": [
          "description",
          "active",
          "address",
          "startChannel",
          "channelCount",
          "id",
          "deDuplicate",
          "priority",
          "monitor",
          "type"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "address": {
            "type": "string",
            "format": "ipv4"
          },
          "channelCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "deDuplicate": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "monitor": {
            "type": "boolean"
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "startChannel": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "type": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      }
    }
  }
}